refactor: Output
use seperate Class for Output
