+++ src/test/java/lotto/PlayTest.java
@@ -47,11 +47,4 @@ class PlayTest {
         assertEquals(62.5, play.profitCalculation(List.of(0, 0, 0, 0, 1, 7)));
         assertEquals(125.0, play.profitCalculation(List.of(0, 0, 0, 0, 2, 6)));
     }
-
-    @Test
-    void printResult() {
-        Play play = new Play();
-        play.printResult(List.of(0, 0, 0, 0, 1, 7));
-    }
-
 }
